refactor(agents): keep bootstrap warning owner reachable
