+++ src/agents/agent-tools.before-tool-call.ts
@@ -18,7 +18,6 @@ export {
   peekAdjustedParamsForToolCall,
 } from "./agent-tools.before-tool-call.state.js";
 export {
-  copyBeforeToolCallHookMarker,
   isToolWrappedWithBeforeToolCallHook,
   setBeforeToolCallDiagnosticsEnabled,
 } from "./before-tool-call-metadata.js";
