+++ src/agents/embedded-agent-runner/run/attempt.spawn-workspace.bootstrap-warning.test.ts
@@ -1,10 +1,10 @@
 // Coverage for bootstrap warning text in system prompt assembly.
 import { describe, expect, it } from "vitest";
+import { buildBootstrapPromptWarning } from "../../bootstrap-budget-warning.js";
 import {
   analyzeBootstrapBudget,
   buildBootstrapPromptWarningNotice,
   buildBootstrapInjectionStats,
-  buildBootstrapPromptWarning,
 } from "../../bootstrap-budget.js";
 import { composeSystemPromptWithHookContext } from "./attempt.thread-helpers.js";
